requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
Django==2.2.3
-e git+https://github.com/cedadev/django-settings-object.git@0288a63cb77bb126fd530438871a0845a05789bc#egg=django_settings_object
djangorestframework==3.10.0
docutils==0.14
dogpile.cache==0.7.1
idna==2.8
iso8601==0.1.12
-e git+https://github.com/cedadev/jasmin-ldap.git@9d62daaeaecca56e320b86f68bee2ad2f745a0ca#egg=jasmin_ldap
jmespath==0.9.4
jsonpatch==1.23
jsonpointer==2.0
keystoneauth1==3.15.0
ldap3==2.6
munch==2.3.2
netifaces==0.10.9
openstacksdk==0.26.0
os-service-types==1.7.0
pbr==5.4.0
psycopg2-binary==2.8.3
pyasn1==0.4.5
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requestsexceptions==1.4.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
urllib3==1.25.3
voluptuous==0.11.5
